Main mostly broken down
