+++ blakearchive/static/directives/scroll-to-top/directive.js
@@ -1,0 +1,13 @@
+angular.module("blake").directive('scrollToTop',function(){
+    var link = function(scope,element,attr) {
+        element.on('click',function(){
+            $('html, body').animate({scrollTop: 0}, 'fast');
+            $("#Overlay").animate({scrollTop: 0}, 'fast');
+        })
+    };
+
+    return{
+        restrict: 'A',
+        link: link
+    }
+});
