+++ blakearchive/static/index.js
@@ -15,6 +15,15 @@ import './js/angular-fullscreen/angular-fullscreen.min';
 import 'script-loader!./js/angular-markdown-it/markdown-it.min';
 import './js/angular-markdown-it/angular-markdown-it';
 import './js/main';
+import './services/generic-service';
+import './services/blake-object';
+import './services/blake-copy';
+import './services/blake-work';
+import './services/blake-data';
+import './services/compare-objects';
+import './services/format';
+import './services/image-manipulation';
+import './services/window-size';
 import './controllers/copy/copy.controller';
 import './controllers/exhibit/exhibit.controller';
 import './controllers/home/home.controller';
@@ -44,3 +53,19 @@ import './directives/text-transcription/textTranscription.directive';
 import './directives/view-sub-menu/viewSubMenu.directive';
 import './directives/work-copies/workCopies.directive';
 import './directives/work-title/workTitle.directive';
+import './directives/magnify-image/directive';
+import './directives/blake-menu/directive';
+import './directives/show-me/directive';
+import './directives/ovp-image/directive';
+import './directives/resize/directive';
+import './directives/auto-height/directive';
+import './directives/affix/directive';
+import './directives/auto-width/directive';
+import './directives/parallax/directive';
+import './directives/left-on-broadcast/directive';
+import './directives/scroll-to-top/directive';
+import './directives/to-top-on-broadcast/directive';
+import './directives/scroll-to-element/directive';
+import './directives/to-top-button/directive';
+import './directives/twitter-share/directive';
+import './filters/highlight';
